[Refactor] macro-configurations-store.cpp/h ファイルを作ってグローバル変数を移した
