+++ src/util/string-processor.h
@@ -7,22 +7,7 @@
 #include <string_view>
 #include <vector>
 
-constexpr size_t MAX_MACRO_MOD = 12;
-constexpr size_t MAX_MACRO_TRIG = 200; /*!< 登録を許すマクロ（トリガー）の最大数 */
-
 extern const char hexsym[16];
-
-enum class ShiftStatus {
-    OFF,
-    ON,
-};
-
-extern size_t max_macrotrigger;
-extern std::optional<std::string> macro_template;
-extern std::optional<std::string> macro_modifier_chr;
-extern std::vector<std::string> macro_modifier_names;
-extern std::vector<std::string> macro_trigger_names;
-extern std::map<ShiftStatus, std::vector<std::string>> macro_trigger_keycodes;
 
 void text_to_ascii(char *buf, std::string_view sv, size_t bufsize);
 void ascii_to_text(char *buf, std::string_view sv, size_t bufsize);
